# docs/specs/swagger.yaml
{"openapi":"3.0.1","info":{"title":"projetofinal","description":"projetofinal","contact":{"name":"projetofinal","email":"ext_dambetan@mercadolibre.com"},"version":"last"},"servers":[{"url":"https://null_projetofinal.furyapps.io","description":"Scope null"}],"paths":{"/api/v1/fresh-products/orders":{"get":{"tags":["purchase-order-controller"],"operationId":"getProductsOfPurchase","parameters":[{"name":"orderId","in":"query","required":true,"schema":{"type":"string"}},{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PurchaseOrderProductResponseDTO"}}}}}}},"put":{"tags":["purchase-order-controller"],"operationId":"update","parameters":[{"name":"orderId","in":"query","required":true,"schema":{"type":"string"}},{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PurchaseOrderResponseDTO"}}}}}},"post":{"tags":["purchase-order-controller"],"operationId":"create_1","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PurchaseValueResponseDTO"}}}}}}},"/api/v1/fresh-products/inboundorder":{"put":{"tags":["insert-order-controller"],"operationId":"update_1","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BatchProductResponseDTO"}}}}}}},"post":{"tags":["insert-order-controller"],"operationId":"create_2","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BatchProductResponseDTO"}}}}}}}},"/api/v1/sign-in":{"post":{"tags":["session-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponseDTO"}}}}}}},"/api/v1/fresh-products/warehouse":{"get":{"tags":["warehouse-controller"],"operationId":"getProductCountByWarehouse","parameters":[{"name":"productId","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductCountListDTO"}}}}}},"post":{"tags":["warehouse-controller"],"operationId":"create","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WarehouseDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WarehouseResponseDTO"}}}}}}},"/api/v1/admin/user":{"get":{"tags":["account-controller"],"operationId":"getUsersByCountry","parameters":[{"name":"country","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountCreatedResponseDTO"}}}}}}},"post":{"tags":["account-controller"],"operationId":"create_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountCreatedResponseDTO"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"operationId":"ping","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/fresh-products":{"get":{"tags":["product-controller"],"operationId":"getAllProducts","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductResponseResponseDTO"}}}}}}}},"/api/v1/fresh-products/representative/list":{"get":{"tags":["batch-product-controller"],"operationId":"findBatchProducts","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}},{"name":"productId","in":"query","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"orderBy","in":"query","required":false,"schema":{"type":"string","default":"nome"}},{"name":"direction","in":"query","required":false,"schema":{"type":"string","default":"DESC"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ListBatchStockSimplifiedResponseDTO"}}}}}}},"/api/v1/fresh-products/list":{"get":{"tags":["product-controller"],"operationId":"getProductsByCategory","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}},{"name":"category","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductResponseResponseDTO"}}}}}}}},"/api/v1/fresh-products/due-date":{"get":{"tags":["batch-product-controller"],"operationId":"findBatchesProductsByDueDate","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}},{"name":"days","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"direction","in":"query","required":false,"schema":{"type":"string","default":"ASC"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BatchStockArrayDueDateResponse"}}}}}}},"/api/v1/fresh-products/due-date/list":{"get":{"tags":["batch-product-controller"],"operationId":"findBatchesProductsByDueDateAndCategory","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}},{"name":"days","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"category","in":"query","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"direction","in":"query","required":false,"schema":{"type":"string","default":"ASC"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BatchStockArrayDueDateResponse"}}}}}}}},"components":{"schemas":{"OrderStatusRequestDTO":{"required":["statusCode"],"type":"object","properties":{"statusCode":{"type":"string","enum":["NEW","RECEIVED","CANCELED","CLOSED"]}}},"ProductsRequestDTO":{"required":["productId","quantity"],"type":"object","properties":{"productId":{"type":"string"},"quantity":{"type":"integer","format":"int32"}}},"PurchaseOrderRequestDTO":{"required":["date"],"type":"object","properties":{"id":{"type":"string"},"date":{"type":"string","format":"date"},"orderStatus":{"$ref":"#/components/schemas/OrderStatusRequestDTO"},"products":{"type":"array","items":{"$ref":"#/components/schemas/ProductsRequestDTO"}}}},"PurchaseRequestDTO":{"type":"object","properties":{"purchaseOrder":{"$ref":"#/components/schemas/PurchaseOrderRequestDTO"}}},"PurchaseOrderProductResponseDTO":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"quantity":{"type":"integer","format":"int32"}}},"PurchaseOrderResponseDTO":{"type":"object","properties":{"id":{"type":"string"},"date":{"type":"string","format":"date"},"orderStatus":{"$ref":"#/components/schemas/OrderStatusRequestDTO"},"products":{"type":"array","items":{"$ref":"#/components/schemas/PurchaseOrderProductResponseDTO"}}}},"BatchProductRequestDTO":{"required":["batchNumber","currentQuantity","currentTemperature","dueDate","initialQuantity","manufacturingDate","manufacturingTime","minimumTemperature","productId"],"type":"object","properties":{"productId":{"type":"string"},"batchNumber":{"type":"integer","format":"int32"},"currentTemperature":{"type":"number","format":"double"},"minimumTemperature":{"type":"number","format":"double"},"initialQuantity":{"type":"integer","format":"int32"},"currentQuantity":{"type":"integer","format":"int32"},"manufacturingDate":{"type":"string","format":"date"},"manufacturingTime":{"type":"string","format":"date-time"},"dueDate":{"type":"string","format":"date"}}},"InboundOrderRequestDTO":{"required":["batchStock","orderDate","orderNumber","section"],"type":"object","properties":{"orderNumber":{"type":"integer","format":"int32"},"orderDate":{"type":"string","format":"date"},"section":{"$ref":"#/components/schemas/SectionOrderRequestDTO"},"batchStock":{"type":"array","items":{"$ref":"#/components/schemas/BatchProductRequestDTO"}}}},"OrderRequestDTO":{"required":["inboundOrder"],"type":"object","properties":{"inboundOrder":{"$ref":"#/components/schemas/InboundOrderRequestDTO"}}},"SectionOrderRequestDTO":{"type":"object","properties":{"sectionCode":{"type":"string"},"warehouseCode":{"type":"string"}}},"BatchProductResponseDTO":{"type":"object","properties":{"batchNumber":{"type":"integer","format":"int32"},"currentTemperature":{"type":"number","format":"double"},"minimumTemperature":{"type":"number","format":"double"},"initialQuantity":{"type":"integer","format":"int32"},"currentQuantity":{"type":"integer","format":"int32"},"manufacturingDate":{"type":"string"},"manufacturingTime":{"type":"string"},"dueDate":{"type":"string"}}},"LoginRequestDTO":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"AccountResponseDTO":{"type":"object","properties":{"username":{"type":"string"},"token":{"type":"string"}}},"WarehouseDTO":{"required":["country","name"],"type":"object","properties":{"name":{"maxLength":2147483647,"minLength":3,"type":"string"},"country":{"maxLength":2147483647,"minLength":3,"type":"string"}}},"WarehouseResponseDTO":{"type":"object","properties":{"message":{"type":"string"},"warehouseDTO":{"$ref":"#/components/schemas/WarehouseDTO"}}},"PurchaseValueResponseDTO":{"type":"object","properties":{"id":{"type":"string"},"totalPrice":{"type":"number","format":"double"}}},"AccountRequestDTO":{"required":["country","password","role","username","warehouseId"],"type":"object","properties":{"username":{"maxLength":50,"minLength":8,"type":"string"},"password":{"pattern":"^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[!@#&()–[{}]:;',?/*~$^+=<>]).{8,20}$","type":"string"},"country":{"type":"string"},"role":{"type":"string"},"warehouseId":{"type":"string"}}},"AccountCreatedResponseDTO":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"warehouse":{"type":"string"},"country":{"type":"string"}}},"ProductResponseResponseDTO":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"price":{"type":"number","format":"double"},"description":{"type":"string"}}},"ProductCountListDTO":{"type":"object","properties":{"productId":{"type":"string"},"warehouses":{"type":"array","items":{"$ref":"#/components/schemas/ProductCountResponseDTO"}}}},"ProductCountResponseDTO":{"type":"object","properties":{"warehouseId":{"type":"string"},"count":{"type":"integer","format":"int32"}}},"BatchStockSimplifiedResponseDTO":{"type":"object","properties":{"batchId":{"type":"string"},"currentQuantity":{"type":"integer","format":"int32"},"dueDate":{"type":"string"}}},"ListBatchStockSimplifiedResponseDTO":{"type":"object","properties":{"section":{"$ref":"#/components/schemas/SectionResponseDTO"},"productId":{"type":"string"},"batchStock":{"type":"array","items":{"$ref":"#/components/schemas/BatchStockSimplifiedResponseDTO"}},"nextPage":{"type":"integer","format":"int32"},"lastPage":{"type":"integer","format":"int32"},"offset":{"type":"integer","format":"int32"}}},"SectionResponseDTO":{"type":"object","properties":{"sectionCode":{"type":"string"},"warehouseCode":{"type":"string"}}},"BatchStockArrayDueDateResponse":{"type":"object","properties":{"batchStock":{"type":"array","items":{"$ref":"#/components/schemas/BatchStockItemDueDateResponse"}},"nextPage":{"type":"integer","format":"int32"},"lastPage":{"type":"integer","format":"int32"},"offset":{"type":"integer","format":"int32"}}},"BatchStockItemDueDateResponse":{"type":"object","properties":{"batchNumber":{"type":"integer","format":"int32"},"productId":{"type":"string"},"productTypeId":{"type":"string"},"dueDate":{"type":"string"},"quantity":{"type":"integer","format":"int32"}}}}}}