# src/main/java/com/mercadolibre/projetofinal/dtos/response/PurchaseValueResponseDTO.java
package com.mercadolibre.projetofinal.dtos.response;

import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
public class PurchaseValueResponseDTO {
    private String id;
    private Double totalPrice;
}